# api/openapi.yml
openapi: 3.0.3
info:
  title: MTS Online Shop
  version: 0.1.0
  description: REST API интернет-магазина
servers:
  - url: http://localhost:8080/api
    description: Dev Server
...
paths:
...
  /cart:
    get:
      summary: Получить содержимое корзины
      tags: [Cart]
      operationId: getCart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Корзина пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
  /cart/items:
    post:
      summary: Добавить товар в корзину
      tags: [Cart]
      operationId: addCartItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCartItemRequest"
      responses:
        "201":
          description: Товар добавлен в корзину
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /cart/items/{itemId}:
    delete:
      summary: Удалить товар из корзины
      tags: [Cart]
      operationId: removeCartItem
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Товар удален из корзины
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /cart/clear:
    delete:
      summary: Очистить корзину
      tags: [Cart]
      operationId: clearCart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Корзина очищена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/users/{userId}/cart:
    get:
      summary: Получить корзину пользователя
      tags: [Admin Users]
      operationId: adminGetUserCart
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Корзина пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
        "401":
          description: Не авторизован
        "403":
          description: Недостаточно прав (только ADMIN)
        "404":
          description: Пользователь не найден
  /admin/users/{userId}/cart/clear:
    delete:
      summary: Очистить корзину пользователя
      tags: [Admin Users]
      operationId: adminClearUserCart
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Корзина пользователя очищена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Не авторизован
        "403":
          description: Недостаточно прав (только ADMIN)
        "404":
          description: Пользователь не найден
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: "Операция выполнена успешно"
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Пакет интернета 5Gb"
        price:
          type: number
          format: double
          example: 199.99
...
    AddCartItemRequest:
      type: object
      required:
        - productId
      properties:
        productId:
          type: integer
          format: int64
          example: 1
    CartItem:
      type: object
      properties:
        itemId:
          type: integer
          format: int64
          example: 123
        product:
          $ref: "#/components/schemas/Product"
    CartResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT